requirements.txt
asgiref==3.8.1
Django==5.2.1
django-cors-headers==4.7.0
django-dotenv==1.4.2
django-filter==25.1
djangorestframework==3.16.0
drf-yasg==1.21.10
inflection==0.5.1
packaging==25.0
pillow==11.2.1
pytz==2025.2
PyYAML==6.0.2
setuptools==72.1.0
sqlparse==0.5.3
uritemplate==4.1.1
wheel==0.43.0
